♻️ refactor: 코드 경로 변경
- application_views.py의 경로를 apps/applications에서 apps/applications/views로 변경
- recruitments_uuid_urls.py와 recruitments_urls.py에 설정된 application_views.py의 경로 수정
