+++ apps/applications/views/application_views.py
